@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace Lsv\TimeHarvestCli\Console\Config;
 
 use Lsv\TimeHarvestCli\Console\AbstractCommand;
-use RuntimeException;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
@@ -31,7 +30,7 @@ class SetDefaultTaskCommand extends AbstractCommand
     {
         // @codeCoverageIgnoreStart
         if (!is_string($input->getOption('task'))) {
-            throw new RuntimeException('Could not parse option "task"');
+            throw new \RuntimeException('Could not parse option "task"');
         }
         // @codeCoverageIgnoreEnd
 
