@@ -4,9 +4,7 @@ declare(strict_types=1);
 
 namespace Lsv\TimeHarvestCli\Console\Lists;
 
-use DateTime;
 use Lsv\TimeHarvestCli\Console\AbstractCommand;
-use RuntimeException;
 use Symfony\Component\Console\Helper\Table;
 use Symfony\Component\Console\Helper\TableSeparator;
 use Symfony\Component\Console\Input\InputArgument;
@@ -25,7 +23,7 @@ class TimeEntryDayCommand extends AbstractCommand
                 'day',
                 InputArgument::OPTIONAL,
                 'The date you want to see, format Y-m-d',
-                (new DateTime())->format('Y-m-d')
+                (new \DateTime())->format('Y-m-d')
             );
     }
 
@@ -33,11 +31,11 @@ class TimeEntryDayCommand extends AbstractCommand
     {
         // @codeCoverageIgnoreStart
         if (!is_string($input->getArgument('day'))) {
-            throw new RuntimeException('Day parameter is not a string');
+            throw new \RuntimeException('Day parameter is not a string');
         }
         // @codeCoverageIgnoreEnd
 
-        $date = new DateTime((string) $input->getArgument('day'));
+        $date = new \DateTime((string) $input->getArgument('day'));
         $entries = $this->client->getTimeEntries($date, $date);
         $json = json_decode($entries, false, 512, JSON_THROW_ON_ERROR);
 
