@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace Lsv\TimeHarvestCli\Console\Time;
 
 use Lsv\TimeHarvestCli\Console\AbstractCommand;
-use RuntimeException;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
@@ -54,7 +53,7 @@ class CreateTimeEntryCommand extends AbstractCommand
 
         // @codeCoverageIgnoreStart
         if (!is_string($input->getOption('project'))) {
-            throw new RuntimeException('Could not parse option "project"');
+            throw new \RuntimeException('Could not parse option "project"');
         }
         // @codeCoverageIgnoreEnd
 
@@ -65,19 +64,19 @@ class CreateTimeEntryCommand extends AbstractCommand
     {
         // @codeCoverageIgnoreStart
         if (!is_string($input->getOption('project'))) {
-            throw new RuntimeException('Could not parse option "project"');
+            throw new \RuntimeException('Could not parse option "project"');
         }
 
         if (!is_string($input->getOption('task'))) {
-            throw new RuntimeException('Could not parse option "task"');
+            throw new \RuntimeException('Could not parse option "task"');
         }
 
         if (!is_string($input->getArgument('hours'))) {
-            throw new RuntimeException('Could not parse option "hours"');
+            throw new \RuntimeException('Could not parse option "hours"');
         }
 
         if (!is_string($input->getArgument('note'))) {
-            throw new RuntimeException('Could not parse option "note"');
+            throw new \RuntimeException('Could not parse option "note"');
         }
         // @codeCoverageIgnoreEnd
 
